[TASK] Add some info about remote control to readme [TASK] Allow setting whisper task in remote control without the tool restarting [TASK] Make osc port and address configurable
