+++ audioWhisper.py
@@ -9,6 +9,7 @@ import click
 import VRC_OSCLib
 import websocket
 import texttranslate
+import settings
 
 
 temp_dir = tempfile.mkdtemp()
@@ -40,9 +41,13 @@ blacklist = list((map(lambda x: x.lower(), blacklist)))
 @click.option("--pause", default=0.8, help="Pause time before entry ends", type=float)
 @click.option("--phrase_time_limit", default=None, help="phrase time limit before entry ends to break up long recognitions.", type=float)
 @click.option("--osc_ip", default="0", help="IP to send OSC message to. Set to '0' to disable", type=str)
+@click.option("--osc_port", default=9000, help="Port to send OSC message to. ('9000' as default for VRChat)", type=int)
+@click.option("--osc_address", default="/chatbox/input", help="The Address the OSC messages are send to. ('/chatbox/input' as default for VRChat)", type=str)
 @click.option("--websocket_ip", default="0", help="IP where Websocket Server listens on. Set to '0' to disable", type=str)
 @click.option("--ai_device", default=None, help="The Device the AI is loaded on. can be 'cuda' or 'cpu'. default does autodetect", type=click.Choice(["cuda", "cpu"]))
-def main(devices, device_index, sample_rate, task, model, english, condition_on_previous_text, verbose, energy, pause,dynamic_energy, phrase_time_limit, osc_ip, websocket_ip, ai_device):
+def main(devices, device_index, sample_rate, task, model, english, condition_on_previous_text, verbose, energy, pause,dynamic_energy, phrase_time_limit, osc_ip, osc_port, osc_address, websocket_ip, ai_device):
+    # set initial settings
+    settings.SetOption("whisper_task", task)
 
     if str2bool(devices) == True:
         index = 0
@@ -75,10 +80,12 @@ def main(devices, device_index, sample_rate, task, model, english, condition_on_
             audio_clip = AudioSegment.from_file(data)
             audio_clip.export(save_path, format="wav")
 
+            whisper_task = settings.GetOption("whisper_task")
+
             if english:
-                result = audio_model.transcribe(save_path, task=task, language='english', condition_on_previous_text=condition_on_previous_text)
+                result = audio_model.transcribe(save_path, task=whisper_task, language='english', condition_on_previous_text=condition_on_previous_text)
             else:
-                result = audio_model.transcribe(save_path, task=task, condition_on_previous_text=condition_on_previous_text)
+                result = audio_model.transcribe(save_path, task=whisper_task, condition_on_previous_text=condition_on_previous_text)
 
             predicted_text = result.get('text').strip()
 
@@ -88,18 +95,18 @@ def main(devices, device_index, sample_rate, task, model, english, condition_on_
                 else:
                     print(result)
                 
-                do_txt_translate = texttranslate.GetOption("txt_translate")
+                do_txt_translate = settings.GetOption("txt_translate")
                 if do_txt_translate:
-                    from_lang = texttranslate.GetOption("src_lang")
-                    to_lang = texttranslate.GetOption("trg_lang")
-                    to_romaji = texttranslate.GetOption("txt_ascii")
+                    from_lang = settings.GetOption("src_lang")
+                    to_lang = settings.GetOption("trg_lang")
+                    to_romaji = settings.GetOption("txt_ascii")
                     predicted_text = texttranslate.TranslateLanguage(predicted_text, from_lang, to_lang, to_romaji)
                     result["txt_translation"] = predicted_text
                     result["txt_translation_target"] = to_lang
 
                 # Send to VRChat
                 if osc_ip != "0":
-                    VRC_OSCLib.Chat(predicted_text, True, "/chatbox/input", IP = osc_ip, PORT = 9000)
+                    VRC_OSCLib.Chat(predicted_text, True, osc_address, IP = osc_ip, PORT = osc_port)
                 # Send to Websocket
                 if websocket_ip != "0":
                     websocket.BroadcastMessage(json.dumps(result))
