+++ websocket.py
@@ -3,6 +3,7 @@ import asyncio
 import websockets
 import json
 import texttranslate
+import settings
 
 WS_CLIENTS = set()
 
@@ -14,14 +15,14 @@ async def handler(websocket):
     await send(websocket, json.dumps({"type": "installed_languages", "data": availableLanguages}))
 
     # send all current text translation settings
-    await send(websocket, json.dumps({"type": "translate_settings", "data": texttranslate.TRANSLATE_SETTINGS}))
+    await send(websocket, json.dumps({"type": "translate_settings", "data": settings.TRANSLATE_SETTINGS}))
 
     WS_CLIENTS.add(websocket)
     try:
         async for message in websocket:
             print("Setting: ", message)
             msgObj = json.loads(message)
-            texttranslate.SetOption(msgObj["name"], msgObj["value"])
+            settings.SetOption(msgObj["name"], msgObj["value"])
             if msgObj["name"] == "dl_langs":
                 texttranslate.InstallLanguages()
         await websocket.wait_closed()
